[refactoring/coupong-22] Leaderboard 관련 4-layered architecture 적용
