+++ src/main/java/com/onepage/coupong/business/leaderboard/LeaderboardService.java
@@ -1,8 +1,9 @@
-package com.onepage.coupong.leaderboard.service;
+package com.onepage.coupong.business.leaderboard;
 
-import com.onepage.coupong.leaderboard.EmitterManager;
+import com.onepage.coupong.implementation.leaderboard.EmitterManager;
+import com.onepage.coupong.presentation.leaderboard.LeaderboardUseCase;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,23 +12,21 @@ import java.util.stream.Collectors;
 
 @Slf4j
 @Service
-public class LeaderboardService {
+@RequiredArgsConstructor
+public class LeaderboardService implements LeaderboardUseCase {
 
-    private final EmitterManager emitterManager; // EmitterManager 주입
-
-    @Autowired
-    public LeaderboardService(EmitterManager emitterManager) {
-        this.emitterManager = emitterManager;
-    }
+    private final EmitterManager emitterManager;
 
     // 리더보드 업데이트
+    @Override
     public void updateLeaderboard(String couponCategory, Set<Object> topWinners, double entryTime) {
         String message = createLeaderboardUpdateMessage(couponCategory, topWinners, entryTime);
         emitterManager.sendMessageToEmitters(message); // EmitterManager를 통해 메시지 전송
     }
 
     // 리더보드 업데이트 메시지 생성
-    private String createLeaderboardUpdateMessage(String couponCategory, Set<Object> topWinners, double entryTime) {
+    @Override
+    public String createLeaderboardUpdateMessage(String couponCategory, Set<Object> topWinners, double entryTime) {
         List<String> winners = topWinners.stream()
                 .map(Object::toString)
                 .collect(Collectors.toList());
