+++ src/main/java/com/onepage/coupong/implementation/leaderboard/EmitterManager.java
@@ -1,4 +1,4 @@
-package com.onepage.coupong.leaderboard;
+package com.onepage.coupong.implementation.leaderboard;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
