+++ src/main/java/com/onepage/coupong/jpa/leaderboard/CouponWinningLog.java
@@ -1,8 +1,8 @@
-package com.onepage.coupong.leaderboard.domain;
+package com.onepage.coupong.jpa.leaderboard;
 
 import com.onepage.coupong.jpa.coupon.Coupon;
 import com.onepage.coupong.user.domain.User;
-import com.onepage.coupong.leaderboard.domain.enums.WinningCouponState;
+import com.onepage.coupong.jpa.leaderboard.enums.WinningCouponState;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
