+++ src/main/java/com/onepage/coupong/jpa/leaderboard/enums/WinningCouponState.java
@@ -1,4 +1,4 @@
-package com.onepage.coupong.leaderboard.domain.enums;
+package com.onepage.coupong.jpa.leaderboard.enums;
 
 public enum WinningCouponState {
     PENDING, COMPLETED, EXPIRED, INVALID
